# openapi/vaultnode.yaml
openapi: 3.1.0
...
info:
  title: VaultNode API
  version: 1.1.0
  summary: Programmatic interface for the VaultNode Turn-Key Filing Protocol
  description: |
    Backend for managing VaultNodes, bridges, mirror checks, versions, witnesses, logs, audits, seals, and refusals.
    Designed for large inline text content (Markdown/YAML) stored in PostgreSQL (TOAST) and accessible by Custom GPTs and web apps.
  contact:
    name: VaultCore Support
    url: https://vaultcore.example.com
servers:
  - url: https://api.vaultcore.example.com/v1
    description: Production (Railway-compatible)
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Static API key for server-to-server/GPT access.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    BridgeIdParam:
      name: bridge_id
      in: path
      required: true
      schema: { type: string }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    OffsetParam:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    VaultNodeInput:
      type: object
      required: [title, author, consent_tier, content]
      properties:
        title: { type: string, minLength: 1 }
        author: { type: string, minLength: 1 }
        contact: { type: string }
        type:
          type: string
          enum: [framework, tool, theory, artifact]
          default: framework
        tags:
          type: array
          items: { type: string }
          default: []
        consent_tier:
          type: string
          enum: [strict, flexible, open]
        description: { type: string }
        provenance: { type: string }
        phase:
          type: string
          enum: [draft, review, sealed]
          default: draft
        metadata:
          type: object
          additionalProperties: true
        resonance_vector:
          type: array
          items: { type: number }
          minItems: 3
          maxItems: 3
        content:
          type: string
          description: Large inline Markdown/YAML/Plaintext
          examples:
            - "# My Framework\n\nThis is the canonical node text..."
    VaultNode:
      allOf:
        - $ref: '#/components/schemas/VaultNodeInput'
        - type: object
          required: [node_id, date_created, status, version]
          properties:
            node_id: { type: string }
            version: { type: string, default: "1.0.0" }
            status: { type: string, default: "active" }
            date_created: { type: string, format: date-time }
            date_sealed: { type: string, format: date-time, nullable: true }
            seal_hash: { type: string, nullable: true }
            witnesses:
              type: array
              items: { $ref: '#/components/schemas/WitnessRef' }
...
    BridgeMapInput:
      type: object
      required: [node_id, target_node, relationship_type]
      properties:
        node_id: { type: string }
        target_node: { type: string }
        relationship_type:
          type: string
          enum: [peer, parent, child, resonance]
        interlink_key: { type: string }
        metrics:
          $ref: '#/components/schemas/BridgeMetrics'
        gate_sequence:
          type: array
          items: { type: string }
          default: []
        status:
          type: string
          enum: [pending, approved, refused]
          default: pending
        decision:
          $ref: '#/components/schemas/Decision'
    BridgeMap:
      allOf:
        - $ref: '#/components/schemas/BridgeMapInput'
        - type: object
          properties:
            id: { type: integer }
            created_at: { type: string, format: date-time }
    BridgeMetrics:
      type: object
      properties:
        sigma: { type: number, minimum: 0 }
        eta: { type: number, minimum: 0, maximum: 1 }
        phi: { type: number, minimum: 0, maximum: 1 }
        delta_hv_pred: { type: number }
        e_deg_risk: { type: number, minimum: 0, maximum: 1 }
    Decision:
      type: object
      properties:
        approved: { type: boolean }
        witness: { type: string }
        timestamp: { type: string, format: date-time }
        notes: { type: string }

    # ---------- Mirrors ----------
...
    SealRequest:
      type: object
      required: [node_id, mode, witness_id]
      properties:
        node_id: { type: string }
        mode:
          type: string
          enum: [soft, hard]
        witness_id: { type: string }
    SealResult:
      type: object
      properties:
        node_id: { type: string }
        mode: { type: string }
        seal_hash: { type: string }
        date_sealed: { type: string, format: date-time }
    ReopenRequest:
      type: object
      required: [node_id, reason, witness_id]
      properties:
        node_id: { type: string }
        reason: { type: string }
        witness_id: { type: string }

    # ---------- Witnesses ----------
...
    WitnessRef:
      type: object
      properties:
        witness_id: { type: string }
        role: { type: string }

    # ---------- Checklists ----------
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
paths:
...
  /bridges:
    post:
      tags: [Bridges]
      summary: Create/Update a BridgeMap between two nodes
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BridgeMapInput' }
      responses:
        '201':
          description: Bridge created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BridgeMap' }
    get:
      tags: [Bridges]
      summary: List bridges (optionally filtered by node_id)
      parameters:
        - name: node_id
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/BridgeMap' }
  /bridges/{bridge_id}:
    get:
      tags: [Bridges]
      summary: Get a bridge
      parameters:
        - $ref: '#/components/parameters/BridgeIdParam'
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BridgeMap' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Bridges]
      summary: Delete a bridge
      parameters:
        - $ref: '#/components/parameters/BridgeIdParam'
      responses:
        '204': { description: Deleted }

  # ===== Mirrors =====
...
  /seal:
    post:
      tags: [Seals]
      summary: Seal a node (soft or hard)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SealRequest' }
      responses:
        '200':
          description: Seal result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SealResult' }
  /reopen:
    post:
      tags: [Seals]
      summary: Reopen a soft-sealed node (with witness)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReopenRequest' }
      responses:
        '200':
          description: Reopened
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VaultNode' }

  # ===== Witnesses =====
...